rename grpc-log to grpclog
